Corrección de errores 1:27

+++ src/main/java/pe/edu/upc/ariseapp/repositories/IRolRepository.java
@@ -1,10 +1,9 @@
 package pe.edu.upc.ariseapp.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import pe.edu.upc.ariseapp.entities.Rol;
 
 @Repository
-public interface IRolRepository extends JpaRepository<Rol, Long> {
+public interface IRolRepository extends JpaRepository<Rol, Integer> {
 }
